Compile on systems that don't export PYTHONPATH
build-all.sh should export PYTHONPATH incase that variable is not
predefined in the environment.

Also fixed: a patch to SConscript in MSYS that so that the compile
is less strict.  Also, note in the README.md that too recent versions
of GCC might not work.  We've tested gcc 4.6.x and 4.7.x.